fix(runtime): refine out-of-process jit ipc
