+++ Cargo.toml
@@ -73,6 +73,9 @@ tracing-subscriber = "0.3"
 tracing-tracy = "0.11"
 paste = "1.0"
 quanta = "0.12"
+serde = { version = "1", features = ["derive"] }
+serde-wincode = "0.1"
+wincode = "0.4.5"
 
 [profile.dev]
 opt-level = 3
