+++ crates/revmc-runtime/src/runtime/config.rs
@@ -1,10 +1,10 @@
 //! Runtime configuration.
 
-use crate::{CompileTimings, runtime::storage::ArtifactStore};
+use crate::{CompileTimings, eyre, runtime::storage::ArtifactStore};
 use alloy_primitives::B256;
 use revm_context_interface::cfg::GasParams;
 use revm_primitives::hardfork::SpecId;
-use std::{env, path::PathBuf, str::FromStr, sync::Arc, time::Duration};
+use std::{path::PathBuf, str::FromStr, sync::Arc, time::Duration};
 
 const JIT_MODE_ENV: &str = "REVMC_JIT_MODE";
 const JIT_HELPER_PATH_ENV: &str = "REVMC_JIT_HELPER_PATH";
@@ -83,15 +83,16 @@ pub struct RuntimeConfig {
 
     /// Where JIT compilation work runs.
     ///
-    /// Defaults to [`JitMode::InProcess`], or `REVMC_JIT_MODE` when set.
+    /// Defaults to [`JitMode::InProcess`].
     pub jit_mode: JitMode,
 
     /// Helper executable used when [`jit_mode`](Self::jit_mode)
     /// is [`JitMode::OutOfProcess`].
     ///
     /// When `None`, the runtime spawns `std::env::current_exe()` and expects it
-    /// to call [`super::maybe_run_jit_helper`] during startup. Defaults to
-    /// `REVMC_JIT_HELPER_PATH` when set.
+    /// to call [`super::maybe_run_jit_helper`] during startup.
+    ///
+    /// Defaults to `None`.
     pub jit_helper_path: Option<PathBuf>,
 
     /// Blocking mode: every lookup synchronously JIT-compiles on miss and never
@@ -158,14 +159,33 @@ impl FromStr for JitMode {
     type Err = String;
 
     fn from_str(s: &str) -> Result<Self, Self::Err> {
-        let mode = s.trim().to_ascii_lowercase();
-        Ok(match mode.as_str() {
-            "" | "in-process" | "in_process" | "inprocess" | "in" | "0" => Self::InProcess,
-            "out-of-process" | "out_of_process" | "outofprocess" | "out" | "oop" | "1" => {
-                Self::OutOfProcess
-            }
+        Ok(match s {
+            "in-process" => Self::InProcess,
+            "out-of-process" => Self::OutOfProcess,
             _ => return Err(format!("unknown JIT mode: {s}")),
         })
+    }
+}
+
+impl RuntimeConfig {
+    /// Applies runtime environment overrides.
+    ///
+    /// Recognized variables are `REVMC_JIT_MODE`, `REVMC_JIT_HELPER_PATH`,
+    /// `REVMC_JIT_HELPER_MEMORY_LIMIT_BYTES`, and `REVMC_JIT_HELPER_CPU_SECONDS`.
+    pub fn with_env_overrides(mut self) -> eyre::Result<Self> {
+        if let Some(mode) = env_var(JIT_MODE_ENV) {
+            self.jit_mode = mode.parse().map_err(|e: String| eyre::eyre!("{JIT_MODE_ENV}: {e}"))?;
+        }
+        if let Some(path) = env_path(JIT_HELPER_PATH_ENV) {
+            self.jit_helper_path = Some(path);
+        }
+        if let Some(limit) = parse_env_u64(JIT_HELPER_MEMORY_LIMIT_ENV)? {
+            self.tuning.jit_helper_memory_limit_bytes = limit;
+        }
+        if let Some(secs) = parse_env_u64(JIT_HELPER_CPU_SECONDS_ENV)? {
+            self.tuning.jit_helper_cpu_time = (secs > 0).then(|| Duration::from_secs(secs));
+        }
+        Ok(self)
     }
 }
 
@@ -182,8 +202,8 @@ impl Default for RuntimeConfig {
             no_dse: false,
             gas_params: None,
             aot: false,
-            jit_mode: env::var(JIT_MODE_ENV).ok().and_then(|s| s.parse().ok()).unwrap_or_default(),
-            jit_helper_path: env_path(JIT_HELPER_PATH_ENV),
+            jit_mode: JitMode::default(),
+            jit_helper_path: None,
             blocking: false,
             on_compilation: None,
         }
@@ -246,7 +266,7 @@ pub struct RuntimeTuning {
     /// `0` disables the limit. On Unix this is applied with `RLIMIT_AS` before
     /// the helper process starts executing.
     ///
-    /// Defaults to `0`, or `REVMC_JIT_HELPER_MEMORY_LIMIT_BYTES` when set.
+    /// Defaults to `0`.
     pub jit_helper_memory_limit_bytes: u64,
 
     /// Maximum CPU time for the out-of-process JIT helper.
@@ -254,8 +274,7 @@ pub struct RuntimeTuning {
     /// `None` disables the limit. On Unix this is applied with `RLIMIT_CPU`
     /// before the helper process starts executing.
     ///
-    /// Defaults to `None`, or `REVMC_JIT_HELPER_CPU_SECONDS` when set to a
-    /// non-zero integer.
+    /// Defaults to `None`.
     pub jit_helper_cpu_time: Option<Duration>,
 
     /// Capacity of the per-worker job queue.
@@ -327,10 +346,8 @@ impl Default for RuntimeTuning {
             jit_max_pending_jobs: 2048,
             jit_worker_count: worker_count,
             jit_timeout: Duration::from_secs(5),
-            jit_helper_memory_limit_bytes: parse_env_u64(JIT_HELPER_MEMORY_LIMIT_ENV).unwrap_or(0),
-            jit_helper_cpu_time: parse_env_u64(JIT_HELPER_CPU_SECONDS_ENV)
-                .filter(|secs| *secs > 0)
-                .map(Duration::from_secs),
+            jit_helper_memory_limit_bytes: 0,
+            jit_helper_cpu_time: None,
             jit_worker_queue_capacity: 64,
             jit_opt_level: crate::OptimizationLevel::default(),
             aot_opt_level: crate::OptimizationLevel::default(),
@@ -342,12 +359,17 @@ impl Default for RuntimeTuning {
     }
 }
 
-fn parse_env_u64(name: &str) -> Option<u64> {
-    env::var(name).ok().and_then(|s| s.parse().ok())
+fn parse_env_u64(name: &str) -> eyre::Result<Option<u64>> {
+    let Some(value) = env_var(name) else { return Ok(None) };
+    value.parse().map(Some).map_err(|e| eyre::eyre!("{name}: {e}"))
+}
+
+fn env_var(name: &str) -> Option<String> {
+    std::env::var(name).ok()
 }
 
 fn env_path(name: &str) -> Option<PathBuf> {
-    env::var_os(name).map(|path| {
+    std::env::var_os(name).map(|path| {
         let path = PathBuf::from(path);
         path.canonicalize().unwrap_or(path)
     })
