+++ docs/out-of-process-jit.md
@@ -24,10 +24,10 @@ Using LLVM ORC's remote executor APIs (`ExecutorProcessControl`, `SimpleRemoteEP
 Current prototype:
 
 - `RuntimeConfig::jit_mode = JitMode::OutOfProcess` makes the runtime keep a global persistent helper process spawned via `RuntimeConfig::jit_helper_path`, or `std::env::current_exe()` when unset.
-- `REVMC_JIT_MODE=out-of-process` switches default runtime configs to out-of-process JIT. `REVMC_JIT_HELPER_PATH` overrides the helper executable path. Test harnesses should point this at a binary that calls `revmc::runtime::maybe_run_jit_helper()` at startup, such as `target/debug/revmc`.
+- `RuntimeConfig::default()` uses plain in-process defaults. `JitBackend::new` applies `RuntimeConfig::with_env_overrides()`, which recognizes `REVMC_JIT_MODE=out-of-process` and `REVMC_JIT_MODE=in-process`; other spellings are rejected. `REVMC_JIT_HELPER_PATH` overrides the helper executable path. Test harnesses should point this at a binary that calls `revmc::runtime::maybe_run_jit_helper()` at startup, such as `target/debug/revmc`.
 - `REVMC_JIT_HELPER_MEMORY_LIMIT_BYTES` and `REVMC_JIT_HELPER_CPU_SECONDS` apply Unix `RLIMIT_AS` and `RLIMIT_CPU` limits to helper processes before `exec`.
 - Helper binaries must call `revmc::runtime::maybe_run_jit_helper()` at process startup. `revmc-cli` does this already.
-- Workers send length-prefixed binary JIT object requests to the helper over stdin and receive length-prefixed binary responses from stdout.
+- Workers send length-prefixed wincode-serialized JIT object requests to the helper over stdin and receive length-prefixed wincode-serialized responses from stdout.
 - The parent links returned object bytes into its local ORC instance, resolves the symbol, and constructs `JitCodeBacking` with a parent-owned `ResourceTracker`.
 - `RuntimeTuning::jit_timeout` bounds each helper compilation; timed-out helpers are killed and replaced on the next job.
 - Clearing resident code or shutting down the runtime kills the helper process so in-flight out-of-process compiles can be interrupted instead of waiting for LLVM to finish.
@@ -42,7 +42,7 @@ Still needed:
 
 ## Open questions
 
-- Serialization crate and stability requirements for the private IPC protocol.
+- Stability requirements for the private IPC protocol.
 - Whether debug dumps should be written by the child, the parent, or both.
 - Whether compiler recycling is still needed per helper worker once the whole helper can be restarted.
 - How to expose helper process configuration such as executable path, environment, and restart policy without making the default API noisy.
